requirements.txt
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
django-shell-plus==1.1.7
djangorestframework==3.14.0
docker==6.0.1
exceptiongroup==1.0.1
executing==1.2.0
flake8==5.0.4
idna==3.4
iniconfig==1.1.1
ipython==8.6.0
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
platformdirs==2.5.3
pluggy==1.0.0
prompt-toolkit==3.0.32
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
pywin32==305
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
tomli==2.0.1
traitlets==5.5.0
tzdata==2022.6
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.2
